VAAPI: fix memory corruption
Update va_surface_ids variable-length array to manually handled pointer.
Storing the address of a local VLA within a pointer which survives the scope
where the decl of that VLA takes place is not a good idea, IMHO. Actually, this
was causing undefined behaviour within the driver (aborting at many random places).