vout-snapshot: udpate vout_snapshot_Set
This patch removes a picture creation and the copy of the pixels and
properties from the source picture (coming from the decoder or video filters)
to this new one, and replaces this behaviour by a simple hold of that source
picture. This fixes the bug that occured when a HW decoder is in use and we're
trying to take a snapshot. The copy described ahead didn't recreate the p_sys
of that picture, which holds the surface id. This resource being absolutly
necessary to retrieve all the image data whatsoever, we must be able to access
the content of the p_sys.