vout: skip unecessary video converters
If the decoder format differs from the one of the display, we instantiate a
filter chain in the display with a video converter to adapt the decoder format
to the display.
If the output format of the last filter of the vout filter chain is the same as
the one of the display but differs from the one of the decoder, then we add a
video converter to match the one of the decoder.
Therefore, we may have unecessary video converters in the case where the last
filter output format matches the one of the display. In this situation, it
would be more efficient to not append a video converter to the vout filter
chain, and just skip the display filter chain.