hw: vaapi-deint: fix incorrect backward/forward reference ordering
After testing the Motion Adaptive deinterlacing mode and finding it produced
lots of flicker/judder and not much deinterlacing I traced the error back to
the below.

Although it seems a little counter-intuitive, the "forward_refs" are actually
the frames older than the current one, and the "backward_refs" are those later.

In addition, the forward_refs (previously the backward_refs) have to be ordered
such that [0] is most recent, [1] older, etc; before this patch it was the
opposite.

Just to be 100% sure, I have also cross-referenced this approach with how
FFmpeg does it:

https://github.com/FFmpeg/FFmpeg/blob/master/libavfilter/vf_deinterlace_vaapi.c

See function deint_vaapi_filter_frame.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>